Fin de Mise en Place de la Gestion des utilisateurs et roles

+++ database/migrations/2020_08_22_152150_create_role_user_table.php
@@ -15,8 +15,8 @@ class CreateRoleUserTable extends Migration
     {
         Schema::create('role_user', function (Blueprint $table) {
             $table->id();
-            $table->integer('role_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->BigInteger('role_id')->unsigned();
+            $table->BigInteger('user_id')->unsigned();
             $table->timestamps();
         });
     }
